requirements.txt
# Основные зависимости для системы email уведомлений

# Коннектор для работы с MySQL базой данных
mysql-connector-python>=8.0.32

# Планировщик задач для автоматического выполнения
APScheduler>=3.10.4

# Загрузка переменных окружения из .env файла
python-dotenv>=1.0.0
